requirements.txt
flask>=2.3.2
youtube-transcript-api>=0.6.1
google-generativeai>=0.3.1,<1.0.0
google-ai-generativelanguage>=0.3.0,<1.0.0
openai>=0.27.8
anthropic>=0.7.0
langdetect>=1.0.9
google-api-core>=2.10.0,<3.0.0
google-api-python-client>=2.80.0,<3.0.0
google-auth>=2.20.0,<3.0.0
google-auth-httplib2>=0.1.0,<1.0.0
annotated-types>=0.7.0,<1.0.0
blinker>=1.8.2,<2.0.0
cachetools>=5.5.0,<6.0.0
certifi>=2024.8.30,<2025.12.31
charset-normalizer>=3.4.0,<4.0.0
click>=8.1.7,<9.0.0
grpcio>=1.50.0,<2.0.0
grpcio-status>=1.50.0,<2.0.0
gunicorn>=21.2.0,<22.0.0
httplib2>=0.22.0,<1.0.0
idna>=3.10,<4.0.0
itsdangerous>=2.2.0,<3.0.0
Jinja2>=3.1.4,<4.0.0
MarkupSafe>=3.0.2,<4.0.0
proto-plus>=1.22.0,<2.0.0
protobuf>=3.19.5,<6.0.0
pyasn1>=0.6.1,<1.0.0
pyasn1_modules>=0.4.1,<1.0.0
pydantic>=2.9.2,<3.0.0
pydantic_core>=2.23.4,<3.0.0
pyparsing>=3.2.0,<4.0.0
requests>=2.32.3,<3.0.0
rsa>=4.9,<5.0.0
six>=1.16.0,<2.0.0
tqdm>=4.66.6,<5.0.0
typing_extensions>=4.12.2,<5.0.0
uritemplate>=4.1.1,<5.0.0
urllib3>=2.2.3,<3.0.0
Werkzeug>=3.1.1,<4.0.0
